requirements.txt
fastapi
uvicorn[standard]
gitpython
chromadb
sentence-transformers
langchain
transformers
accelerate
torch
pygithub
flake8
pylint
autopep8
python-dotenv
pydantic
requests
jinja2
python-multipart
